+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Props `noOptionText` to `QasAutoComplete`.
+- Added the possibility to pass slots from the quasar select to our `QasAutoComplete` component.
+- Added type `search-select` in `QasField` component.
+
 ## 2.8.0 - 2021-08-27
 
 ### Added